Created the transaction engine.
A user can now send a transaction to the endpoint /transaction-engine with a POST request and corresponding payload. The transaction engine will then check inventory per a given store and then determine if the items are available. If at any point the inventory deduction fails, the transaction engine will place the items back. If the inventory substraction is successful, the transaction engine will then insert row into the transaction details table. Once the transaction details have been implemented, the total transaction cost will be calculated and the corresponding sales tax determined where this store is located. The transaction engine is self-correcting and will clean up any tables as needed sending a response of a successful or unsuccessful transaction with a corresponding payload.